Added support for optional backtracking in parser library
The motivation for this change is better error reporting. Before
this commit, running `alt p1 p2` would try to parse `p1` and upon
failure, parse the same input with `p2`. However, if `p1` has
started consuming input before it fails, backtracking hides errors
in subexpressions. This commit removes this behaviour and
introduces a combinator `try` which backtracks on failure. The old
behaviour can be restored with `alt (try p1) p2`.