Memoize generated symbols based on types
Whenever a new constructor is defined, instead of blindly
generating a new symbol, this commit makes it so that the symbol
for each combination of constructor and datatype is remembered and
reused. This means that two datatypes with the same names are
assumed to be "the same".

Two obvious TODOs:
- Names (and thereby, symbols) should not be resolved during
runtime, but via static analysis (i.e., typechecking).
- We currently ignore parameters of datatypes and look only at the
rightmost type in a chain of arrows.