Fix bug in intersection between edge patterns + more tests
If we are taking the intersection between two edge patterns, where
one pattern has a longer prefix than the other, and the other has
a longer postfix than the other, then the intersection should contain
some exact length lists that are shorter than (max pre1 pre2) +
(max post1 post2), namely those where the extraneous patterns overlap.
For example, [!1] ++ _ & _ ++ [!2] is the same as
[!1] ++ _ ++ [!2] | [!(1 | 2)], where the second part of the union would
not show up in the old implementation.