Tup build system (#674)
* Make mi-lite take the output filename as an argument

* Tup build system

Create configuration to build and test everything using Tup. Most of the interesting stuff happens in `Tuprules.lua`, which defines how to build, run, and test normal `.mc` files, as well as a list of exceptions, files that shouldn't be built/evaled. Bootstrapping is handled by the root `Tupfile`.

Much like the make-based build system there is a list of files we don't expect to compile/run/etc. However, we make sure to *try* compiling/running/etc. anyway; we just check that the exit code is non-zero. This means that we can have negative tests, and that the lists have to be up-to-date.

We're also discovering *all* `.mc` files in the repo, as opposed to only those in folders we've explicitly marked for inclusion. For files where the tests are too slow to run regularly there are two lists that avoid running/evaling altogether: `noRun` and `noEval`.

Worth looking into, maybe for later updates:
- Non-standard compilation tests (tuning, accelerate, other backends, etc.)
- Cleverness regarding installed dependencies, it's currently set according to what succeeds on my machine.
- I tried listing *why* things failed in the exclusion lists, and some of the failures seem strange.