packaging: guix: Depropagate inputs.
This style of packaging is more in line with Guix practice.  We also add some
basic native inputs that will anyhow be required, but probably aren't
sufficient for the purpose of cross-builds.  Since currently only
"test-compile" is used for testing however, this might actually be enough.

We also split gcc-toolchain, since it is usually understood that a user brings
all of their compilers when packaging or writing a build system.