Packaging updates (#759)
* packaging: guix: Use git-file-name for dependencies.

* packaging: guix: Build from git sources.

* packaging: guix: Depropagate inputs.

This style of packaging is more in line with Guix practice.  We also add some
basic native inputs that will anyhow be required, but probably aren't
sufficient for the purpose of cross-builds.  Since currently only
"test-compile" is used for testing however, this might actually be enough.

We also split gcc-toolchain, since it is usually understood that a user brings
all of their compilers when packaging or writing a build system.

* packaging: nix: Build from git sources.

* stdlib: sys: Use POSIX-compliant `command' instead of `which'.

* stdlib: sys: Add maximum number of tries for sysTempMake

* packaging: guix: Let the user provide OCaml and Dune.

This is in line with the treatment of gcc-toolchain, and makes things
a bit easier when making a binary tarball with miking-pack, since
opam's OCaml and dune are easier to relocate.

* packaging: guix: Use later commit for ocaml-ISO8601.

* packaging: guix: Remove ocaml-base-bytes.

This package is unnecessary provided ocaml-findlib is available

* packaging: guix: Use dune-build-system instead of gnu-build-system.

This enables writing, e.g., (package-with-ocaml5.0 miking).

* packaging: guix: Replace variable ocamlprefix with ocamllibdir.

* packaging: nix: Update to match the guix definition.

* packaging: nix: Use HEAD ref in fetchGit source.

This makes each new commit trigger a rebuild, rather than every
change.

* packaging: Update README.md

---------

Co-authored-by: Liliana Marie Prikler <liliana.prikler@ist.tugraz.at>